+++ Automate/Framework/AutomationFactory.cs
@@ -162,8 +162,8 @@ namespace Pathoschild.Stardew.Automate.Framework
             // machine
             if (feature is FruitTree fruitTree)
                 return new FruitTreeMachine(fruitTree, location, tile);
-            if (feature is Bush bush && bush.size.Value == Bush.greenTeaBush && !bush.townBush.Value)
-                return new TeaBushMachine(bush, location, tile);
+            if (feature is Bush bush && BushMachine.CanAutomate(bush))
+                return new BushMachine(bush, location);
 
             // connector
             if (this.IsConnector(feature))